# openapi.yaml
openapi: "3.1.1"
info:
  title: Rave Collection E-commerce API
  description: API for Rave Collection shoe shop e-commerce and inventory management system
  version: 1.0.0
  contact:
    name: Rave Collection
    email: support@ravecollection.com
servers:
  - url: http://localhost:5005/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get all orders (Admin only)
      responses:
        "200":
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      tags:
        - Orders
      summary: Create new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}/status:
    put:
      tags:
        - Orders
      summary: Update order status (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, completed, rejected]
      responses:
        "200":
          description: Order status updated

  # Billing endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [customer, admin]
        created_at:
          type: string
          format: date-time
    UserRegistration:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [customer, admin]
          default: customer
...
    Order:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        customer_name:
          type: string
        total_amount:
          type: number
          format: decimal
        status:
          type: string
          enum: [pending, completed, rejected]
        created_at:
          type: string
          format: date-time
    OrderInput:
      type: object
      required:
        - items
        - total_amount
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
              quantity:
                type: integer
              price:
                type: number
                format: decimal
        total_amount:
          type: number
          format: decimal
...